Moved structure expectations into __tests__ to exclude from codecov
