+++ src/__tests__/Fields.spec.js
@@ -9,9 +9,9 @@ import createReducer from '../createReducer'
 import createFields from '../createFields'
 import FormSection from '../FormSection'
 import plain from '../structure/plain'
-import plainExpectations from '../structure/plain/expectations'
+import plainExpectations from '../structure/plain/__tests__/expectations'
 import immutable from '../structure/immutable'
-import immutableExpectations from '../structure/immutable/expectations'
+import immutableExpectations from '../structure/immutable/__tests__/expectations'
 
 
 const describeFields = (name, structure, combineReducers, setup) => {
