+++ src/__tests__/Form.spec.js
@@ -9,9 +9,9 @@ import createReducer from '../createReducer'
 import createField from '../createField'
 import Form from '../Form'
 import plain from '../structure/plain'
-import plainExpectations from '../structure/plain/expectations'
+import plainExpectations from '../structure/plain/__tests__/expectations'
 import immutable from '../structure/immutable'
-import immutableExpectations from '../structure/immutable/expectations'
+import immutableExpectations from '../structure/immutable/__tests__/expectations'
 
 import SubmissionError from '../SubmissionError'
 import actions from '../actions'
