+++ src/__tests__/FormSection.spec.js
@@ -11,9 +11,9 @@ import createFields from '../createFields'
 import createFieldArray from '../createFieldArray'
 import FormSection from '../FormSection'
 import plain from '../structure/plain'
-import plainExpectations from '../structure/plain/expectations'
+import plainExpectations from '../structure/plain/__tests__/expectations'
 import immutable from '../structure/immutable'
-import immutableExpectations from '../structure/immutable/expectations'
+import immutableExpectations from '../structure/immutable/__tests__/expectations'
 
 const describeFormSection = (name, structure, combineReducers, setup) => {
   const reduxForm = createReduxForm(structure)
