+++ src/__tests__/formValues.spec.js
@@ -11,9 +11,9 @@ import formValues from '../formValues'
 import immutableFormValues from '../immutable/formValues'
 import FormSection from '../FormSection'
 import plain from '../structure/plain'
-import plainExpectations from '../structure/plain/expectations'
+import plainExpectations from '../structure/plain/__tests__/expectations'
 import immutable from '../structure/immutable'
-import immutableExpectations from '../structure/immutable/expectations'
+import immutableExpectations from '../structure/immutable/__tests__/expectations'
 
 
 const describeValues = (
