+++ src/__tests__/generateValidator.spec.js
@@ -1,8 +1,8 @@
 import generateValidator from '../generateValidator'
 import plain from '../structure/plain'
-import plainExpectations from '../structure/plain/expectations'
+import plainExpectations from '../structure/plain/__tests__/expectations'
 import immutable from '../structure/immutable'
-import immutableExpectations from '../structure/immutable/expectations'
+import immutableExpectations from '../structure/immutable/__tests__/expectations'
 
 
 const describeGenerateValidator = (name, structure, setup) => {
