+++ src/__tests__/reduxForm.spec.js
@@ -12,9 +12,9 @@ import createReducer from '../createReducer'
 import createReduxForm from '../createReduxForm'
 import FormSection from '../FormSection'
 import immutable from '../structure/immutable'
-import immutableExpectations from '../structure/immutable/expectations'
+import immutableExpectations from '../structure/immutable/__tests__/expectations'
 import plain from '../structure/plain'
-import plainExpectations from '../structure/plain/expectations'
+import plainExpectations from '../structure/plain/__tests__/expectations'
 import SubmissionError from '../SubmissionError'
 
 import FormWrapper from '../Form'
