+++ src/__tests__/values.spec.js
@@ -7,9 +7,9 @@ import TestUtils from 'react-dom/test-utils'
 import createReducer from '../createReducer'
 import createValues from '../createValues'
 import plain from '../structure/plain'
-import plainExpectations from '../structure/plain/expectations'
+import plainExpectations from '../structure/plain/__tests__/expectations'
 import immutable from '../structure/immutable'
-import immutableExpectations from '../structure/immutable/expectations'
+import immutableExpectations from '../structure/immutable/__tests__/expectations'
 
 
 const describeValues = (name, structure, combineReducers, setup) => {
