+++ src/selectors/__tests__/isValid.spec.js
@@ -1,8 +1,8 @@
 import createIsValid from '../isValid'
 import plain from '../../structure/plain'
-import plainExpectations from '../../structure/plain/expectations'
+import plainExpectations from '../../structure/plain/__tests__/expectations'
 import immutable from '../../structure/immutable'
-import immutableExpectations from '../../structure/immutable/expectations'
+import immutableExpectations from '../../structure/immutable/__tests__/expectations'
 
 
 const describeIsValid = (name, structure, setup) => {
